Further work on movement and weapons system
I have continued to fix problems with the weapons script and changing the project document to reflect these changes. I decided to focus on the weapons system and not work on AI due to a lack of time. I am having trouble adding vertical view to camera as it results in glitching.